changed serial from 8 to 32 bit

+++ birdwing_codegen/contexts/machine_specific_toolhead_structs.py
@@ -84,7 +84,7 @@ base_structs_dict = {
                 {"name": "extrusion_distance_mm", "type": "uint32",
                  "default": "0"},
                 {"name": "refurb_count", "type": "uint8", "default": "0"},
-                {"name": "serial", "type": "uint8", "default": "0"}
+                {"name": "serial", "type": "uint32", "default": "0"}
             ]
         }
     ]
